rag/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
aiosqlite==0.20.0
pandas==2.2.2
openpyxl==3.1.2
pdfplumber==0.11.0
anthropic==0.25.0
openai==1.3.5
python-multipart==0.0.9
pydantic==2.7.1
python-jose==3.3.0
passlib==1.7.4
httpx==0.27.0
python-dotenv==1.0.0
pillow==10.1.0
reportlab==4.0.7
matplotlib==3.8.2
numpy==1.26.2